repository: Move object repacking into housekeeping package
We want to allow for OptimizeRepository to be called without invoking an
RPC in order to allow us to call it in more contexts. As a preparatory
step, we thus have to move all functionality which is invoked by it into
a package that is independent of gRPC services.

Move the logic handling repacking of objects into the housekeeping
package as part of this move.