migrations: Flatten replication job JSON
We are gonna to add foreign key constraints to
the replication_queue table to auto-remove any
rows that become orphan (no repository exists).
Because repository_id is a field inside the job
column of type JSONB we can't create a foreign
key on it. Those we need to create a column for that
field. Because we already know the full stable
set of the columns we need for the replication job
we extract all fields from the job column to be
just simple columns of the table. It is expected
to have a non-null values for repository_id,
replica_path and source_node_storage columns for
the existing rows as we don't handle NULLs in go
code.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3102