coordinator: Add tests for gRPC errors to determine outdated nodes
The tests which determine whether `getUpdatedAndOutdatedNodes()` works
correctly is using standard Go errors to verify whether comparison of
errors works as expected. This is not the typical kind of errors we'd
get there though, but instead we'd usually get gRPC errors which cannot
be directly compared with each other.

Add some testcases for gRPC-style errors and record their current
behaviour.