limithandler: fix flaky test TestLimitHandlerMetrics
In this test, we didn't wait for both requests to return. Thus the test
was ending and the grpc server was being shut down before the request
could return.

Fix this by putting both requests onto the request channel and waiting
for both requests to be received before exiting the test.