praefect: Pass feature flags to Gitaly when executing ServerInfo
When Praefect receives a ServerInfo RPC it forwards the request to
proxied Gitaly nodes to obtain their information and then pass it
through back to the caller. We do not forward the RPC metadata correctly
though, and as a result any feature flags passed to Praefect will not
end up with Gitaly.

Fix this by converting the incoming to an outgoing context before doing
the RPC calls to Gitaly.

Changelog: fixed