housekeeping: Create new manager to host stateful logic
We're eventually going to move the housekeeping package to keep state
such that we can improve the heuristics used for optimizing repos. This
will allow us to e.g. take into account the last time a repository was
optimized to not optimize repositories too often, or avoid packing the
same repository twice concurrently.

Implment a new housekeeping manager structure that will be the home of
this functionality and inject it as required. This manager isn't yet
doing anything.