git: Remove support for the Ruby hooks directory
Historically, our Git hooks were located in the Ruby directory as a set
of wrapper scripts which derive the actual location of the gitaly-hooks
binary via the `GITALY_BIN_DIR` environment variable. This has been
replaced with a new mode though, where we instead create a temporary
directory which contains symlinks for all of the hooks which directly
point to gitaly-hooks. This is done both to remove the indirection via
another shell script, and more importantly to get rid of our dependency
on the Ruby directory.

This new mode has been rolled out as default-enabled in 14.8, so we can
now remove support for the old Ruby hooks directory in this release.

Changelog: removed