cmd/gitaly-hooks: Modernize tests to execute gitaly-hooks directly
Tests for the gitaly-hooks binary are still using the Ruby wrapper
scripts. These are deprecated though, where the new way to execute the
binary is instead via a temporary directory created at runtime which
only contains symlinks to the binary.

Refactor the tests to execute the binary directly now that we're about
to remove the Ruby hook scripts. This also requires us to move to the
new way to derive which hook is about to be executed: the hook name is
not derived via the first argument anymore, but instead via the zeroth
argument which is set to the name of the symlink.