cmd/gitaly-hooks: Remove compatibility code to resolve hook names
When we introduced the hook symlinks which directly point to the
gitaly-hooks binary we also had to adapt gitaly-hooks such that it can
continue to resolve the hook name it ought to execute. Previously, this
name was derived via the first argument, which was passed to us via a
wrapper script. This was not possible anymore though given that there
is no wrapper script anymore. Instead, gitaly-hooks gained support for
resolving the hook name via its zeroth argument, which is the name of
the symlink we're executing and thus contains the hook name, as well.
We had to retain backwards-compatibility with the old way of resolving
the hook name though to not break hook executions during zero-downtime
upgrades.

Now that the old code is gone though we do not need the compatibility
code anymore, so this commit removes it.