test-boot: Write gitlab-shell secret into temporary directory
The test-boot script verifies that Gitaly comes up correctly and is used
as a very basic smoke test in our CI. As part of the bringup it writes
the gitlab-shell secret into the `ruby/git-hooks` directory such that
Gitaly is able to find the secret, or otherwise it would fail to boot.

This directory is going away, so we need to find a better location. And
we do in fact already have one: we can just write it into a subdirectory
of the temporary directory which already contains all the other runtime
data required for a successful boot. Convert the script to do so.