reconciler: Use new columns instead of JBONB job column
As we migrate JSONB job column to a set of simple columns
in this commit we change query that references deprecated
job column with references on the corresponding simple
columns of the replication_queue table.

Because tests were written with some inconsistency in this
commit we address it by providing a proper value for the
Change filed of the ReplicationJob where it is missed.
And because of the foreign key constraint between
replication_queue and repositories table in it not possible
anymore to create replication events for non-existing
repositories. That is why setup code of some test-case
was changed.