migration: Indexes re-creation on replication_queue
As we flattened job column of JSONB type to a set of
simple columns we deprecate usage of the job column
and replaces it with references on the corresponding
columns of the replication_queue table for all the
indexes that are exist on the replication_queue. Now
they are referencing a new columns.