Move GetReplicaPath to gittest
This commit moves GetReplicaPath helper function to gittest package.
This groups the function better with other git repository related
helpers and helps avoid a cyclic dependency between gittest and
testhelper package when a later commit adds more helpers.

The helper is also changed to take in a config instead of ClientConn.
Many contexts where the helper is used don't have the ClientConn
readily available but they all have a config available. This improves
the ergonomics of the test helper.