datastore: Use new columns instead of JBONB job column
As we flattened job column of JSONB type to a set of
simple columns we deprecate usage of the job column
and replaces it with references on the corresponding
columns of the replication_queue table for
PostgresReplicationEventQueue.
The data is retrieved from the columns on the scan
operation. The tests are fixed to include repository_id
into replication jobs.