Rewrite repository in NewTestRepo
This commit rewrites the repository in NewTestRepo helper. NewTestRepo
instantiates a new localrepo from a Repository proto. Since all of the
call sites are passing in a repository that has not yet been rewritten,
the git operations target wrong directories when the tests are ran with
Praefect in front. Rewriting the repository ensures the operations target
the replica path generated by Praefect rather than the relative path sent
by the client.

As TestWriteCommit in gittest depends on NewTestRepo, this change would
introduce a cyclic dependency between the localrepo and gittest packages.
The test is moved into the external _test package to break this cycle.