Prepare object pool service's tests for replica path
Object pool service's test are working on the disk directly and they
assume the object pool is stored at the relative path the client sent.
This won't be the case anymore with Praefect rewriting relative paths
to guarantee atomicity of creations, deletes and renames.

This commit adjusts the tests to access the object pool at the path
generated by Praefect if the tests are ran with Praefect in front.