Prepare repository service's tests for replica path
Repository service's test are working on the disk directly and they
assume the repositories are stored at the relative path the client sent.
This won't be the case anymore with Praefect rewriting relative paths
to guarantee atomicity of creations, deletes and renames. This commit
updates the tests to access the repositories via the replica path
generated by Praefect if Praefect is enabled in the tests.