Prepare smarthttp service's tests for replica path
smarthttp service's test are working on the disk directly and they
assume the object pool is stored at the relative path the client sent.
This won't be the case anymore with Praefect rewriting relative paths
to guarantee atomicity of creations, deletes and renames.

This commit adjusts the tests to account for the relative path rewriting
by accessing the repositories on the disk via the Praefect generated
replica path if Praefect is enabled in the tests. Likewise, the cache keys
in the tests have been updated to take into account the rewritten relative
path.