Prepare operation service's tests for replica path
TestApplyPatch is currently building its test state by accessing the
repository at its relative path. With Praefect rewriting the relative
paths, this is not the case if Praefect is enabled in the tests.
This commit updates the tests to account for relative path rewriting
so the git operations end up targeting the correct directory.