Prepare remote service's tests for replica path
TestUpdateRemoteMirror sets up the state in the test repository by
accessing it via its relative path. This commit updates the test to
access the repository at it's rewritten location if Praefect is ran in
front of the test.

While at it, the test cases are marked with t.Parallel() to speed up the
test's execution.