Update quarantine repository tests for relative path rewriting
Quarantine directories are set in Gitaly to receive objects into a
separate directory from the repository's main object database. The
tests are not currently account for relative path rewriting that
Praefect would do. To do so, this commit changes the tests to
set the quarantine directories based on the rewritten relative path.
The API still needs to be called with the relative path so Praefect
can route the request correctly.