git: Simplify setup of pack-generation configuration
We use a flag to tag all Git commands which may potentially generate a
packfile, where this information is then used to inject specific config
into these commands which changes the way the packfiles are generated.
This is needlessly roundabout: we can just append this configuration to
the per-command options.

Convert the code to do this to make it easier to understand.