proto: Deprecate fine-grained repository maintenance RPCs
Historically we had and still have lots of fine-grained RPCs which
perform repository maintenance in Gitaly. This meant that control of how
repositories are packed does not reside in Gitaly though, and this is
starting to become a bigger problem because we cannot easily iterate on
how repositories should be maintained. We have thus recently extended
the black-box-style maintenance RPC OptimizeRepository with the intent
of making it the single source of truth for how repositories are packed.

Deprecate the old fine-grained repository maintenance RPCs. They will be
removed as soon as all callers have been migrated to OptimizeRepository,
which will then allow us to improve the general repository maintenance
strategy.

Changelog: deprecated