repository: Ignore missing refs in CreateBundleFromRefList
This RPC is used to create backup bundles. The list of refs passed in is
collected by a call to ListRefs. If a ref is deleted between these two
calls then the `git bundle create` call made by CreateBundleFromRefList
would fail with an error.

We are not interested in missing refs for backups because there will be
no need to restore a now deleted ref. However, ignoring missing refs
will mean that the list of refs saved for incremental backups will be
out-of-date compared to the bundle. There should be no problem creating
incremental backups because we only need the object ID that the ref used
to be pointed at. So the ref names would only be used to detect ref
deletions from the previous backup on restore
(https://gitlab.com/gitlab-org/gitaly/-/issues/3857), any extra refs in
the ref list can simply be ignored.

Changelog: fixed