praefect: Repository should exist before scheduling events
After introduction of the repository ID we use it
for various operations where the relative path and
virtual storage were used. One of the use cases is
creation of the replication events. The injected
repository ID is used to identify concrete repository
later. Because now we have a foreign key we are not
allowed to create replication events for repositories
that are not exist yet in the database. That is why
we use PostgresRepositoryStore to create a repository
records before we enqueue any replication events.

The change also contains fixes of some inconsistencies
in setup logic as double-database creation, unneeded
transactions usage.