Remove a Praefect specific backup restoration test
The backup package has a test for restoring a backup through Praefect.
The test manually sets up a Praefect in front of a Gitaly. This is not
necessary as the `testserver.RunGitalyServer` automatically sets up a
Praefect when the `GITALY_TEST_WITH_PRAEFECT` env is set. As such, the
test case currently just duplicates the `TestManager_Restore` case. This
commit removes the case and collapses the `testManagerRestore` helper
into the `TestManager_Restore` test case.