Pull unconfigured socket path message into a constant
testcfg.Build doesn't yet know the socket path when building the
configuration. The socket path is substituted into the configuration
when the service being spun up. The message used to bypass config
validation error is spread in multiple locations in the code base.
This commit pulls it into a constant.