gitaly: Fix missing housekeeping manager dependency
While we set up the housekeeping manager when starting Gitaly, we don't
inject it into the service dependencies. As a result, Gitaly will not
have it set up during normal runtime operations and will thus trigger a
panic whenever trying to access it. This went unnoticed because our test
setup for the Gitaly server is different from the setup we use when
running tests.

Fix this bug by injecting the housekeeping manager as expected

Changelog: fixed