backchannel: do not block when closing client
In https://gitlab.com/gitlab-org/gitlab/-/issues/351425 we learned
that when grpc-go reconnects to a Gitaly server, it synchronously
closes the old connection. That makes it important that we do not
block while closing a connection. This commit removes code that waited
for the backchannel client-side server to close while closing a client
connection. Instead of waiting, we just propagate the close to the
yamux sesion.

Changelog: fixed