git: Track supported execution environments in a single place
The knowledge about which execution environments exist is distributed
across multiple different locations: first we need to set them up, and
second we need to make use them depending on whether they have been
configured and enabled. This is really fragile and ultimately hard to
maintain.

We now have the logic to set up the different environments in self
contained structures which all have the same interface. It is thus
trivial to move this logic into a single place, which is a new array of
all supported execution environment constructors. The Git command
factory can thus be stripped of all knowledge around which different
execution environments exists and only needs to loop through all
constructors -- if at least one of them returns a valid environment,
then we're good to go.

We need to maintain two properties though:

    - We prioritize the order in which we consult the different
      environments. If bundled Git is configured, then it should be
      preferred over distributed Git. And if distributed Git is
      configured, it should be preferred over the fallback Git
      environment which was resolved via PATH.

    - Environments must be feature-flaggable. So even if an environment
      with a higher priority is available, it may still be disabled
      which would thus cause us to pick a lower-priority backend.

These properties are maintained by encoding the priority of backends via
the order in which environment constructors are inside their array. The
Git command factory thus constructs all environments in the order they
are listed in and will then loop through them. The first backend that is
enabled (that is, all its feature flags are turned on) will be picked.

Like this, it is trivial to add new execution environments as they are
all tracked in a single central location.