Test pruneIfNeeded through the public API
TestPruneIfNeeded tests that the repository gets pruned if needed. More
importantly, it verifies that object pools do not get pruned. Right now
the tests does not create the repositories through the API. This means
that the test won't catch problems if the pruning code in Gitaly doesn't
recognize the object pool paths generated by Praefect. This commit moves
the test into the external test package so it can run a test server without
cyclic dependency issues. Doing so, the test is changed to test the functionality
through the public API of the housekeeping package. This improves the tests
coverage while still being able to assert the pruning through the log
messages.

The substests are marked as parallel while at it to improve the run time
slightly.