git: Reject config with bundled Git and missing binary directories
When bundled Git is configured then the binaries are expected to be
contained in Gitaly's binary directory. It is thus mandatory for the
directory to be set, or else the resulting execution environment cannot
actually be a valid one. And while we already reject the case where the
binary directory is unset in case the `GITALY_TESTING_BUNDLED_GIT_PATH`
environment variable is set, we don't have such a test for "normal"
setups.

Add the missing sanify check.

Changelog: fixed