git: Refactor CheckRefFormat
CheckRefFormat calls git-check-ref-format(1) and if there is an error
from the command, returns a separate CheckRefFormat error. This doesn't
help much however. It swallows up the context of the original error. But
also, the implementation of git-check-ref-format(1) is such that if the
format is not valid, it will return with an exit code of 1.

Clean up the interface of this function by instead, checking if the
exit code of git-check-ref-format(1) is 1. If it is, then we know the
format is invalid. If the exit code is anything else, something went
wrong so we bubble up the error to be handled by the caller. We can also
get rid of CheckRefFormatError.

Update the one callsite that checks for CheckRefFormatError