operations: Return error from UserSquash when merging fails
We do not return an error in case creating the squashed merge commit in
UserSquash call fails. This makes us blind for the error conditions in
this RPC, which is in turn creating problems in Praefect setups where we
have to rely on errors in order to decide whether we have to create
replication jobs or not.

Return an error in case the merge fails. Note that we don't return a
structured error in this case: the range of commits we're about to
squash-merge has already been rebased, so we do not expect to ever see a
merge conflict here. As a consequence, a failure to create the merge
commit is unexpected and thus doesn't warrant its own error type.

Changelog: changed