git: Verify that bundled Git binaries exist and are executable
We do not verify that all expected bundled Git binaries exist and are
executable when they have been configured by the administrator. As a
result, Gitaly will happly start up but then later fail at runtime when
trying to use the bundled Git execution environment in case any of the
binaries is missing.

Tighten our sanity checks to verify that the bundled Git environment is
set up correctly.

Changelog: fixed