localrepo: Add test to demonstrate failure to unlock symrefs
We have recently added transactional support for writing symbolic
references in a repository via the `safe.LockingFileWriter` interface.
We forgot to unlock the writer in all cases though, which means that in
some error conditions we may leave behind a lockfile.

Add a testcase which demonstrates this failure to unlock.