operations: Return error from UserSquash when resolving revisions fails
We do not return an error in case resolving either the start or end
revision of a UserSquash call fails. This makes us blind for the error
conditions in this RPC, which is in turn creating problems in Praefect
setups where we have to rely on errors in order to decide whether we
have to create replication jobs or not.

Return structured errors in case the error condition triggers such that
callers can tell what kind of error they have been hitting.

Changelog: changed