repository: Allow CreateRepository to take default_branch
In GitLab rails, there is code to ensure that the default branch on a
new repository is set. This is currently done with an extra WriteRef
call. This is done in a racy way, which causes problems with transaction
voting. To fix this, we can simplify the logic altogether  and avoid
adding another roundtrip, add a parameter in CreateRepositoryRequest
to pass a default_branch.

Changelog: changed