middleware/cache: Handle maintenance-style RPCs
The caching middleware is responsible for the invalidation of the disk
cache when it receives an RPC that modifies the on-disk data. This
behaviour is implemented open-coded: only if we receive an RPC that is
either not repository-scoped or that is an accessor will we not
invalidate the repository. As a result, any newly added operation types
will fall back to invalidating the cache.

The newly added maintenance operations will right now also default to
invalidating any caches. This behaviour indeed matches previous their
previous behaviour where they are classified as mutators. But in fact it
is not necessary to invalidate any caches when performing maintenance
tasks: they should be idempotent with regards to the user-visible state,
so the result of any accessor should not change if such an RPC was
executed in between.

Extend the middleware cache to not invalidate disk caches when receiving
a maintenance operation.

Changelog: performance