praefect: Implement routing for maintenance operations
We have introduced a new "maintenance" operation type for RPCs which
allows us to easily special-case repositories which fall into this
class. One those locations where we do wish to special-case them is in
the coordinator. All RPCs which we're about to convert to be maintenance
operations are currently classified as mutators. As a consequence, we:

    - Need to explicitly special-case those RPCs to not have
      transactions enabled because they shouldn't change any on-disk
      state as visible to the user. Furthermore, the state we're writing
      to disk is often not deterministic.

    - Because transactions are disabled, maintenance is never performed
      on multiple nodes at once. As a consequence, we always create
      replication jobs for maintenance jobs, which requires quite some
      architecture and persistence of parameters in the database.

    - We skip nodes which are not currently consistent with the state
      other nodes have. This is not a requirement though: because the
      maintenance RPCs shouldn't have an impact on user-visible state
      anyway we don't really care about consistency, only about
      routability.

Introduce special routing for those new maintenance operations. We're
using a best-effort strategy: every node that is hosting the repository
at hand and which is online will be included. Furthermore, we never
create replication jobs. It should be acceptable to just wait for the
next time a maintenance RPC is scheduled because they shouldn't have an
impact on servicability.

With these changes in place we can eventually get rid of replication for
maintenance-style RPCs, which is a good step towards deprecating the
complete replication queue.

Changelog: changed