praefect: Revert relative path rewriting in RepositoryReplicas
With 8d3767c71 (Rewrite relative path in RepositoryReplicas, 2022-03-02)
we have introduced rewriting of the repository's relative path for the
RepositoryReplicas RPC. This has introduced new usage of the repository
store into the code path, which is required such that we can obtain the
replica path from the database. Unfortunately, this is causing failures
downstream in Rails, which is still not using a database and thus uses a
stubbed-out repository store.

Fixing this isn't quite trivial because we can't create a simple stub
implementation of `GetReplicaPath()`. We'd have two choices: either
return successfully with a bogus replica path, or return an error like
`ErrRepositoryNotFound`, and neither of those options fixes the failure
observed in Rails.

Revert the commit for now. We're actively working to migrate Rails to
use a database for its tests, so the limitations we have will go away.
Furthermore, the actual path rewriting is blocked by the very same
migration, so the revert shouldn't block any ongoing work either.